requirements.txt
boto3==1.28.41
botocore==1.31.41
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
jmespath==1.0.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
tqdm==4.66.1
urllib3>=2.0.4
